@@ -1,20 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "com.bignerdranch.android.mockwalker"
-        minSdkVersion 16
-        targetSdkVersion 25
+        minSdkVersion 24
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        //butterknife
-        javaCompileOptions {
-            annotationProcessorOptions {
-                includeCompileClasspath = true
-            }
-        }
     }
     buildTypes {
         release {
@@ -26,8 +20,8 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:25.3.0'
-    implementation 'com.google.android.gms:play-services-location:9.8.0'
-    implementation 'com.jakewharton:butterknife:6.1.0'
-    implementation 'io.reactivex:rxjava:1.0.8'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.gms:play-services-location:17.0.0'
+    implementation 'io.reactivex:rxjava:1.3.8'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 }
