@@ -17,6 +17,6 @@ public class App extends Application {
     public void onCreate() {
         super.onCreate();
         instance = this;
-        StrictMode.enableDefaults();
+//        StrictMode.enableDefaults();
     }
 }
